requirements.txt
tensorflow-gpu
Pillow
scipy==1.0.0